==> copy_paste_n_json/00_make_nodsize_csv.ipynb <==
from glob import glob
import os
import numpy as np
import pandas as pd

## Make nod_size csv (with smoothing)

### All nod

with open('../preprocess/nod_2925.csv', 'r') as f:
    df = pd.read_csv(f)
    
df.rename(columns={'path':'nod_path', 'nii_path':'img_path'},inplace=True)
    
img_nii_path = df['img_path'].values
nod_path = df['nod_path'].values

for idx in range(len(df_nod)):
    img_nii_path[idx] = img_nii_path[idx].replace('NII_norm', 'NII_normwinall')
    nod_path[idx] = nod_path[idx].replace('NII_norm', 'NII_normwinall')

df.to_csv('../nod_smoothing.csv', index=False)

### nod > 10

with open('../nod_smoothing_10.csv', 'r') as f:
    df = pd.read_csv(f)
df_nod=df[(df.height >= 10) & (df.width >= 10)].reset_index(drop=True)
df_nod.to_csv('../nod_smoothing_10_10mm.csv', index=False)

### nod > 20

with open('../nod_smoothing_10.csv', 'r') as f:
    df = pd.read_csv(f)
df_nod=df[(df.height >= 20) & (df.width >= 20)].reset_index(drop=True)
df_nod.to_csv('../nod_smoothing_10_20mm.csv', index=False)
