==> copy_paste_n_json/06_opencxr.ipynb <==
import opencxr
from opencxr.utils.file_io import read_file, write_file
from opencxr.utils import apply_size_changes_to_img

%matplotlib inline
import matplotlib.pyplot as plt
import matplotlib.patches as patches

import os
import numpy as np
import pandas as pd

from tqdm import tqdm
from glob import glob

import datetime
from PIL import Image

import random

cxr_std_algorithm = opencxr.load(opencxr.algorithms.cxr_standardize)

## Convert train data

drr_path = '../data/PNG_normwinall_train_20mm'
mask_path = '../data/DRR_mask'
save_all_path = '../data/only_20_original/data_all'
save_mask_path = '../data/only_20_original/mask_all'
save_train_path = '../data/only_20_original/train'
save_valid_path = '../data/only_20_original/valid'
save_test_path = '../data/only_20_original/test'

train_neg = pd.read_csv('../train_neg_revised.csv')
train_neg.head(5)

img_path_lst = train_neg.img_path.values
check_double = []
for img_path in tqdm(img_path_lst):
    if img_path in check_double:
        continue
    else:
        check_double.append(img_path)
        file_name = img_path.split('/')[-1].replace('nii.gz', 'png')
        img_np, spacing, _ = read_file(os.path.join(drr_path, file_name))
        std_img, new_spacing, size_changes = cxr_std_algorithm.run(img_np, spacing)
        std_img = np.transpose(std_img)
        
        plt.imsave(os.path.join(save_all_path, file_name), std_img, cmap='Greys_r')
        plt.imsave(os.path.join(save_train_path, file_name), std_img, cmap='Greys_r')

train_pos = pd.read_csv('../train_pos_revised.csv')

check_double = []
for _, row in tqdm(train_pos.iterrows()):
    if row.img_path in check_double:
        pass
    else:
        check_double.append(row.img_path)
        file_name = row.img_path.split('/')[-1].replace('nii.gz', 'png')
        img_np, spacing, _ = read_file(os.path.join(drr_path, file_name))
        std_img, new_spacing, size_changes = cxr_std_algorithm.run(img_np, spacing)
        std_img = np.transpose(std_img)
        
        plt.imsave(os.path.join(save_all_path, file_name), std_img, cmap='Greys_r')
        plt.imsave(os.path.join(save_train_path, file_name), std_img, cmap='Greys_r')
    
    mask_file_name = row.nod_path.split('/')[-1].replace('nii.gz', 'png')
    img_np_mask, spacing_mask, _ = read_file(os.path.join(drr_path, mask_file_name))
    new_mask_png, _ = apply_size_changes_to_img(img_np_mask, new_spacing, size_changes)
    new_mask_png = np.transpose(new_mask_png)
    
    plt.imsave(os.path.join(save_mask_path, mask_file_name), new_mask_png, cmap='Greys_r')

## Convert valid data

valid_neg = pd.read_csv('../valid_neg_revised.csv')

img_path_lst = valid_neg.img_path.values
check_double = []
for img_path in tqdm(img_path_lst):
    if img_path in check_double:
        continue
    else:
        check_double.append(img_path)
        file_name = img_path.split('/')[-1].replace('nii.gz', 'png')
        img_np, spacing, _ = read_file(os.path.join(drr_path, file_name))
        std_img, new_spacing, size_changes = cxr_std_algorithm.run(img_np, spacing)
        std_img = np.transpose(std_img)
        
        plt.imsave(os.path.join(save_all_path, file_name), std_img, cmap='Greys_r')
        plt.imsave(os.path.join(save_valid_path, file_name), std_img, cmap='Greys_r')

valid_pos = pd.read_csv('../valid_pos_revised.csv')

check_double = []
for _, row in tqdm(valid_pos.iterrows()):
    if row.img_path in check_double:
        pass
    else:
        check_double.append(row.img_path)
        file_name = row.img_path.split('/')[-1].replace('nii.gz', 'png')
        img_np, spacing, _ = read_file(os.path.join(drr_path, file_name))
        std_img, new_spacing, size_changes = cxr_std_algorithm.run(img_np, spacing)
        std_img = np.transpose(std_img)
        
        plt.imsave(os.path.join(save_all_path, file_name), std_img, cmap='Greys_r')
        plt.imsave(os.path.join(save_valid_path, file_name), std_img, cmap='Greys_r')
    
    mask_file_name = row.nod_path.split('/')[-1].replace('nii.gz', 'png')
    img_np_mask, spacing_mask, _ = read_file(os.path.join(drr_path, mask_file_name))
    new_mask_png, _ = apply_size_changes_to_img(img_np_mask, new_spacing, size_changes)
    new_mask_png = np.transpose(new_mask_png)
    
    plt.imsave(os.path.join(save_mask_path, mask_file_name), new_mask_png, cmap='Greys_r')

## Convert test data

test_pos = pd.read_csv('../test_pos_revised.csv')

check_double = []
for _, row in tqdm(test_pos.iterrows()):
    if row.img_path in check_double:
        pass
    else:
        check_double.append(row.img_path)
        file_name = row.img_path.split('/')[-1].replace('nii.gz', 'png')
        img_np, spacing, _ = read_file(os.path.join(drr_path, file_name))
        std_img, new_spacing, size_changes = cxr_std_algorithm.run(img_np, spacing)
        std_img = np.transpose(std_img)
        
        plt.imsave(os.path.join(save_all_path, file_name), std_img, cmap='Greys_r')
        plt.imsave(os.path.join(save_test_path, file_name), std_img, cmap='Greys_r')
    
    mask_file_name = row.nod_path.split('/')[-1].replace('nii.gz', 'png')
    img_np_mask, spacing_mask, _ = read_file(os.path.join(drr_path, mask_file_name))
    new_mask_png, _ = apply_size_changes_to_img(img_np_mask, new_spacing, size_changes)
    new_mask_png = np.transpose(new_mask_png)
    
    plt.imsave(os.path.join(save_mask_path, mask_file_name), new_mask_png, cmap='Greys_r')
